# perf(ext/event): optimize addEventListener options converter