@marcosc90 2023-08-18
> Unfortunately, this is not correct. For example, passing a `boolean` as the third argument to `Event` should throw, but now doesn't.

You mean `.addEventlistener('foo', null, true)` because I tested that (also with all other types) and it does not throw on `main`.