@marcosc90 2023-08-18
> Unfortunately, this is not correct. For example, passing a `boolean` as the third argument to `Event` should throw, but now doesn't.

I tested that and it does not throw on main.


> I'd urge you to look at the implementation of `createDictionaryConverter`. This output is very general because it has to deal with any possible dictionary, and a hand-written converter can definitely be faster, but you need to take significant care to not regress the correctness.

Please provide an input that is not correct, because run several manual tests on Chrome/Node/Firefox as well as my implementation and couldn't find any difference on behavior. Also all WPT pass correctly. Once I have the incorrect value for options that my patch is not handling correctly I'll do the adjustments.