@lucacasonato 2023-08-18
@marcosc90 For the prototype access bug - does it regress performance if you branch on the `webidl.type(V) !== "Object"` like this:

```
if (webidl.type(V) !== "Object") return { capture: true, passive: false, once: false, signal: undefined };
const options = {
  capture: !!V.capture,
  passive: !!V.passive,
  once: !!V.once,
};
...

return options;
```