@lucacasonato 2023-08-18
Great, then I think that should fix that bug.